app/requirements.txt
pollination-streamlit>=0.5.0
streamlit==1.23.1
pandas==1.4.1
honeybee-core==1.55.15
honeybee-energy==1.100.1
pollination-streamlit-io==0.71.1
